more methods in testenv
