+++ examples/nft/src/lib.rs
@@ -4,16 +4,15 @@ use near_contract_standards::non_fungible_token::metadata::{
 use near_contract_standards::non_fungible_token::NonFungibleToken;
 use near_contract_standards::non_fungible_token::{Token, TokenId};
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
-use near_sdk::collections::LazyOption;
 use near_sdk::{
-    env, near_bindgen, AccountId, BorshStorageKey, PanicOnDefault, Promise, PromiseOrValue, base64,
+    env, near_bindgen, AccountId, BorshStorageKey, PanicOnDefault, Promise, PromiseOrValue,
 };
 use quickjs_rust_near::web4::types::{Web4Request, Web4Response};
-use quickjs_rust_near::jslib;
+use quickjs_rust_near::jslib::{js_get_property, js_get_string, load_js_bytecode, js_call_function};
 mod web4content;
 use web4content::{INDEX_HTML, MUSIC_WASM, SERVICEWORKER};
+use std::ffi::CStr;
 
-const DATA_IMAGE_SVG_NEAR_ICON: &str = "data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 288 288'%3E%3Cg id='l' data-name='l'%3E%3Cpath d='M187.58,79.81l-30.1,44.69a3.2,3.2,0,0,0,4.75,4.2L191.86,103a1.2,1.2,0,0,1,2,.91v80.46a1.2,1.2,0,0,1-2.12.77L102.18,77.93A15.35,15.35,0,0,0,90.47,72.5H87.34A15.34,15.34,0,0,0,72,87.84V201.16A15.34,15.34,0,0,0,87.34,216.5h0a15.35,15.35,0,0,0,13.08-7.31l30.1-44.69a3.2,3.2,0,0,0-4.75-4.2L96.14,186a1.2,1.2,0,0,1-2-.91V104.61a1.2,1.2,0,0,1,2.12-.77l89.55,107.23a15.35,15.35,0,0,0,11.71,5.43h3.13A15.34,15.34,0,0,0,216,201.16V87.84A15.34,15.34,0,0,0,200.66,72.5h0A15.35,15.35,0,0,0,187.58,79.81Z'/%3E%3C/g%3E%3C/svg%3E";
 static QUICKJS_BINARY: &'static [u8] = include_bytes!("quickjsbytecode.bin");
 
 #[derive(BorshSerialize, BorshStorageKey)]
@@ -82,16 +81,36 @@ near_contract_standards::impl_non_fungible_token_enumeration!(Contract, tokens);
 #[near_bindgen]
 impl NonFungibleTokenMetadataProvider for Contract {
     fn nft_metadata(&self) -> NFTContractMetadata {
-        let val = jslib::run_js_bytecode(QUICKJS_BINARY.to_vec());
+        let jsmod = load_js_bytecode(QUICKJS_BINARY.to_vec());
         
-        NFTContractMetadata {
-            spec: NFT_METADATA_SPEC.to_string(),
-            name: "Example NEAR non-fungible token".to_string(),
-            symbol: "EXAMPLE".to_string(),
-            icon: Some(DATA_IMAGE_SVG_NEAR_ICON.to_string()),
-            base_uri: None,
-            reference: None,
-            reference_hash: None,
+        unsafe {
+            let val = js_call_function(jsmod, "nft_metadata".as_ptr() as i32);
+            let name = CStr::from_ptr(js_get_string(js_get_property(val, "name".as_ptr() as i32)) as *const i8).to_str().unwrap();
+            let symbol = CStr::from_ptr(js_get_string(js_get_property(val, "symbol".as_ptr() as i32)) as *const i8).to_str().unwrap();
+            let icon = CStr::from_ptr(js_get_string(js_get_property(val, "icon".as_ptr() as i32)) as *const i8).to_str().unwrap();
+
+            NFTContractMetadata {
+                spec: NFT_METADATA_SPEC.to_string(),
+                name: name.to_string(),
+                symbol: symbol.to_string(),
+                icon: Some(icon.to_string()),
+                base_uri: None,
+                reference: None,
+                reference_hash: None,
+            }
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use quickjs_rust_near::testenv::testenv::{alice, set_signer_account_id, setup_test_env, assert_latest_return_value_string_eq};
+
+    #[test]
+    fn test_nft_metadata() {
+        setup_test_env();
+        let contract = Contract::new();
+        contract.nft_metadata();
     }
 }
